optimize initial document load
instead of setting the Document via slate-automerge-doc, which in turn
uses a normal slate editing operation to load the initial document,
use the initial value facility of slate instead.
This seems to cut down the time spent to get the initial document loaded
into slate by a factor of 4 to 5.